+++ examples/custom_quad/src/main.rs
@@ -200,6 +200,7 @@ mod quad {
                         radius: self.radius,
                         width: self.border_width,
                         color: Color::from_rgb(1.0, 0.0, 0.0),
+                        ..Border::default()
                     },
                     shadow: self.shadow,
                     snap: self.snap,