+++ widget/src/toggler.rs
@@ -435,6 +435,7 @@ where
                     radius: border_radius,
                     width: style.background_border_width,
                     color: style.background_border_color,
+                    ..Border::default()
                 },
                 ..renderer::Quad::default()
             },
@@ -461,6 +462,7 @@ where
                     radius: border_radius,
                     width: style.foreground_border_width,
                     color: style.foreground_border_color,
+                    ..Border::default()
                 },
                 ..renderer::Quad::default()
             },